Update README.md
Paul the Apostle - A movie based on the apostle's Paul last days bringing inspiration in faith. 

Interstellar - A drama sci-fi movie that portraits the beauty of space and science, importance of family and time. 

Warcraft - A movie based on videogame. Exploring the lore and beauty of the world building of Warcraft.